Add the tokenizer feature and entrypoint
This feature allow us to tokenize code dynamically by using our CLI.

This is a tiny utility to allow us to design a more flexible syntax
parser with well know and pre-defined templates to retrieve functions
definitions, exception throwing, etc.. just by passing a template of
tokens to catch. This feature aim to help us to design this.